feat(effects): Implement Effects Engine
