+++ src/main.cpp
@@ -6,6 +6,7 @@
 #include "SettingsManager.h"
 #include "ButtonManager.h"
 #include "MatrixDisplayManager.h"
+#include "EffectsEngine.h"
 
 // ===============================================
 // CONFIGURATION & CONSTANTS
@@ -27,43 +28,6 @@ uint8_t oePin      = 15;
 #define TEXT_SIZE_MIN 1
 #define TEXT_SIZE_MAX 3
 #define BRIGHTNESS_LEVELS 10
-
-// Confetti Settings
-#define NUM_CONFETTI   40
-#define CONFETTI_RAD   1
-
-// Matrix Rain Settings
-#define NUM_MATRIX_DROPS 12
-#define MATRIX_CHAR_DELAY 80
-
-// Torrent Settings
-#define NUM_TORRENT_DROPS 30
-#define TORRENT_CHAR_DELAY 60
-
-// Star Settings
-#define NUM_STARS 25
-#define STAR_TWINKLE_CHANCE 5
-
-// Shooting Star Settings
-#define NUM_SHOOTING_STARS 2
-#define SHOOTING_STAR_SPEED 0.8f
-#define SHOOTING_STAR_TRAIL_LENGTH 8
-
-// Sparkle Settings
-#define NUM_SPARKLES 30
-#define SPARKLE_DURATION 800
-
-// Fireworks Settings
-#define NUM_FIREWORKS 8
-#define FIREWORK_PARTICLES 15
-#define FIREWORK_LIFE 1500
-
-// Tron Settings
-#define NUM_TRON_TRAILS 12
-#define TRON_MIN_LENGTH 8
-#define TRON_MAX_LENGTH 20
-#define TRON_MIN_SPEED 80
-#define TRON_MAX_SPEED 200
 
 // Timing Constants
 #define MENU_DELAY 30               // Reduced from 50ms 
@@ -92,62 +56,6 @@ enum SetClockStep {
   CONFIRM 
 };
 
-struct Confetti { 
-  float x, y, vx, vy; 
-  uint16_t color; 
-};
-
-struct MatrixDrop {
-  float x, y;
-  uint8_t length;
-  uint8_t speed;
-  uint32_t lastUpdate;
-};
-
-struct Star {
-  float x, y;  // Current position (can be off-screen)
-  uint8_t brightness;
-  uint8_t twinkleState;
-  uint32_t lastTwinkle;
-};
-
-struct ShootingStar {
-  float x, y;
-  float speedX, speedY;
-  bool active;
-  uint8_t trailLength;
-  uint32_t spawnTime;
-};
-
-struct Sparkle {
-  uint8_t x, y;
-  uint8_t brightness;
-  uint16_t color;
-  uint32_t startTime;
-  uint16_t duration;
-};
-
-struct Firework {
-  float x, y;
-  float vx[FIREWORK_PARTICLES], vy[FIREWORK_PARTICLES];
-  uint8_t px[FIREWORK_PARTICLES], py[FIREWORK_PARTICLES];
-  uint16_t color;
-  uint32_t startTime;
-  bool active;
-  bool exploded;
-};
-
-struct TronTrail {
-  uint8_t x, y;
-  uint8_t direction;  // 0=right, 1=down, 2=left, 3=up
-  uint8_t trailPositions[TRON_MAX_LENGTH][2];  // Store trail segment positions
-  uint8_t currentLength;
-  uint16_t color;
-  uint32_t lastMove;
-  uint16_t speed;
-  bool active;
-};
-
 // ===============================================
 // GLOBAL VARIABLES
 // ===============================================
@@ -160,6 +68,7 @@ RTC_DS3231 rtc;
 SettingsManager settings;
 ButtonManager buttons;
 MatrixDisplayManager display(&matrix, &settings);
+EffectsEngine effects(&display, &settings);
 
 // Display Settings
 uint16_t textColors[BRIGHTNESS_LEVELS] = { 0x2104, 0x4208, 0x630C, 0x8410, 0xA514, 0xC618, 0xE71C, 0xEF5D, 0xF79E, 0xFFFF };
@@ -194,21 +103,6 @@ uint32_t lastEnterPress = 0;  // Track last ENTER button press
 const uint32_t ENTER_COOLDOWN = 300;  // 300ms cooldown between ENTER presses
 bool entryLockProcessed = false;  // Track if entry lock has been processed
 char timeStr[16];
-
-// Effects
-Confetti confetti[NUM_CONFETTI];
-MatrixDrop matrixDrops[NUM_MATRIX_DROPS];
-MatrixDrop torrentDrops[NUM_TORRENT_DROPS];
-Star stars[NUM_STARS];
-ShootingStar shootingStars[NUM_SHOOTING_STARS];
-Sparkle sparkles[NUM_SPARKLES];
-Firework fireworks[NUM_FIREWORKS];
-TronTrail tronTrails[NUM_TRON_TRAILS];
-
-// Shooting star timing variables
-uint32_t lastShootingStarTime = 0;
-bool waitingForSecondStar = false;
-uint32_t secondStarTimer = 0;
 
 // ===============================================
 // UTILITY FUNCTIONS
@@ -512,46 +406,8 @@ float generateVelocity(float minSpeed, float maxSpeed, bool allowNegative = true
 /**
  * Initialize confetti particles
  */
-void initializeConfetti() {
-  int16_t centerY = MATRIX_HEIGHT / 2;
-  int16_t yRange = 12;
-  
-  for (int i = 0; i < NUM_CONFETTI; i++) {
-    confetti[i].x = random(0, MATRIX_WIDTH);
-    confetti[i].y = random(centerY - yRange, centerY + yRange);
-    
-    // Ensure minimum velocity to prevent stuck particles
-    confetti[i].vx = generateVelocity(0.1, 0.8);  // Min 0.1, max 0.8 pixels/frame
-    confetti[i].vy = generateVelocity(0.05, 0.4); // Min 0.05, max 0.4 pixels/frame
-    
-    confetti[i].color = randomVividColor();
-  }
-}
 
-/**
- * Reset a confetti particle to a new random position
- */
-void resetConfettiParticle(int index) {
-  int16_t centerY = MATRIX_HEIGHT / 2;
-  int16_t yRange = 12;
-  
-  // Spawn from edges for more dynamic movement
-  if (random(2) == 0) {
-    // Spawn from left or right edge
-    confetti[index].x = random(2) == 0 ? -CONFETTI_RAD : MATRIX_WIDTH + CONFETTI_RAD;
-    confetti[index].y = random(centerY - yRange, centerY + yRange);
-  } else {
-    // Spawn from top or bottom edge
-    confetti[index].x = random(0, MATRIX_WIDTH);
-    confetti[index].y = random(2) == 0 ? -CONFETTI_RAD : MATRIX_HEIGHT + CONFETTI_RAD;
-  }
-  
-  // Ensure minimum velocity to prevent stuck particles
-  confetti[index].vx = generateVelocity(0.1, 0.8);
-  confetti[index].vy = generateVelocity(0.05, 0.4);
-  
-  confetti[index].color = randomVividColor();
-}
+
 
 /**
  * Get the bounding box of the current time display (main clock only)
@@ -617,626 +473,7 @@ bool isInTextArea(int x, int y) {
   return false;
 }
 
-/**
- * Update and draw confetti animation with text masking
- */
-void updateConfetti() {
-  // First pass: update positions
-  for (int i = 0; i < NUM_CONFETTI; i++) {
-    // Update position
-    confetti[i].x += confetti[i].vx;
-    confetti[i].y += confetti[i].vy;
-    
-    // Reset if particle goes off screen
-    if (confetti[i].x < -CONFETTI_RAD * 2 || confetti[i].x > MATRIX_WIDTH + CONFETTI_RAD * 2 ||
-        confetti[i].y < -CONFETTI_RAD * 2 || confetti[i].y > MATRIX_HEIGHT + CONFETTI_RAD * 2) {
-      resetConfettiParticle(i);
-    }
-    
-    // If particle has been moving too slowly for too long, reset it
-    if (abs(confetti[i].vx) < 0.05 && abs(confetti[i].vy) < 0.05) {
-      resetConfettiParticle(i);
-    }
-  }
-  
-  // Second pass: draw particles that are not behind text
-  for (int i = 0; i < NUM_CONFETTI; i++) {
-    int px = (int)confetti[i].x;
-    int py = (int)confetti[i].y;
-    
-    // Only draw if not in text area
-    if (!isInTextArea(px, py)) {
-      uint16_t brightColor = applyEffectBrightness(confetti[i].color);
-      matrix.fillCircle(px, py, CONFETTI_RAD, brightColor);
-    }
-  }
-}
 
-/**
- * Initialize matrix rain drops
- */
-void initializeMatrixRain() {
-  for (int i = 0; i < NUM_MATRIX_DROPS; i++) {
-    matrixDrops[i].x = random(0, MATRIX_WIDTH);
-    matrixDrops[i].y = random(-20, 0);  // Start above screen
-    matrixDrops[i].length = random(3, 8);
-    matrixDrops[i].speed = random(1, 4);
-    matrixDrops[i].lastUpdate = millis();
-  }
-}
-
-/**
- * Update matrix rain effect
- */
-void updateMatrixRain() {
-  uint32_t currentTime = millis();
-  
-  for (int i = 0; i < NUM_MATRIX_DROPS; i++) {
-    if (currentTime - matrixDrops[i].lastUpdate > MATRIX_CHAR_DELAY / matrixDrops[i].speed) {
-      matrixDrops[i].y += matrixDrops[i].speed;
-      matrixDrops[i].lastUpdate = currentTime;
-      
-      // Reset when drop goes off screen
-      if (matrixDrops[i].y > MATRIX_HEIGHT + matrixDrops[i].length) {
-        matrixDrops[i].x = random(0, MATRIX_WIDTH);
-        matrixDrops[i].y = random(-20, -5);
-        matrixDrops[i].length = random(3, 8);
-        matrixDrops[i].speed = random(1, 4);
-      }
-    }
-    
-    // Draw the drop trail
-    for (int j = 0; j < matrixDrops[i].length && (matrixDrops[i].y - j) >= 0; j++) {
-      int y = matrixDrops[i].y - j;
-      if (y < MATRIX_HEIGHT && !isInTextArea(matrixDrops[i].x, y)) {
-        uint8_t intensity = 255 - (j * 40);  // Fade as we go up
-        if (intensity < 50) intensity = 50;
-        uint16_t color = scaledEffectColor565(0, intensity, 0);  // Green rain with brightness scaling
-        matrix.drawPixel(matrixDrops[i].x, y, color);
-      }
-    }
-  }
-}
-
-/**
- * Initialize rain (blue raindrops)
- */
-void initializeRain() {
-  for (int i = 0; i < NUM_MATRIX_DROPS; i++) {
-    matrixDrops[i].x = random(0, MATRIX_WIDTH);
-    matrixDrops[i].y = random(-20, 0);  // Start above screen
-    matrixDrops[i].length = random(3, 8);
-    matrixDrops[i].speed = random(1, 4);
-    matrixDrops[i].lastUpdate = millis();
-  }
-}
-
-/**
- * Update rain effect (blue raindrops)
- */
-void updateRain() {
-  uint32_t currentTime = millis();
-  
-  for (int i = 0; i < NUM_MATRIX_DROPS; i++) {
-    if (currentTime - matrixDrops[i].lastUpdate > MATRIX_CHAR_DELAY / matrixDrops[i].speed) {
-      matrixDrops[i].y += matrixDrops[i].speed;
-      matrixDrops[i].lastUpdate = currentTime;
-      
-      // Reset when drop goes off screen
-      if (matrixDrops[i].y > MATRIX_HEIGHT + matrixDrops[i].length) {
-        matrixDrops[i].x = random(0, MATRIX_WIDTH);
-        matrixDrops[i].y = random(-20, -5);
-        matrixDrops[i].length = random(3, 8);
-        matrixDrops[i].speed = random(1, 4);
-      }
-    }
-    
-    // Draw the drop trail in blue
-    for (int j = 0; j < matrixDrops[i].length && (matrixDrops[i].y - j) >= 0; j++) {
-      int y = matrixDrops[i].y - j;
-      if (y < MATRIX_HEIGHT && !isInTextArea(matrixDrops[i].x, y)) {
-        uint8_t intensity = 255 - (j * 40);  // Fade as we go up
-        if (intensity < 50) intensity = 50;
-        uint16_t color = scaledEffectColor565(0, 0, intensity);  // Blue rain with brightness scaling
-        matrix.drawPixel(matrixDrops[i].x, y, color);
-      }
-    }
-  }
-}
-
-/**
- * Initialize torrent (many small raindrops)
- */
-void initializeTorrent() {
-  for (int i = 0; i < NUM_TORRENT_DROPS; i++) {
-    torrentDrops[i].x = random(0, MATRIX_WIDTH);
-    torrentDrops[i].y = random(-30, 0);  // Start above screen
-    torrentDrops[i].length = random(1, 4);  // Smaller drops
-    torrentDrops[i].speed = random(2, 6);   // Faster speed
-    torrentDrops[i].lastUpdate = millis();
-  }
-}
-
-/**
- * Update torrent effect (many small fast raindrops)
- */
-void updateTorrent() {
-  uint32_t currentTime = millis();
-  
-  for (int i = 0; i < NUM_TORRENT_DROPS; i++) {
-    if (currentTime - torrentDrops[i].lastUpdate > TORRENT_CHAR_DELAY / torrentDrops[i].speed) {
-      torrentDrops[i].y += torrentDrops[i].speed;
-      torrentDrops[i].lastUpdate = currentTime;
-      
-      // Reset when drop goes off screen
-      if (torrentDrops[i].y > MATRIX_HEIGHT + torrentDrops[i].length) {
-        torrentDrops[i].x = random(0, MATRIX_WIDTH);
-        torrentDrops[i].y = random(-30, -5);
-        torrentDrops[i].length = random(1, 4);  // Smaller drops
-        torrentDrops[i].speed = random(2, 6);   // Faster speed
-      }
-    }
-    
-    // Draw the drop trail in white/light blue
-    for (int j = 0; j < torrentDrops[i].length && (torrentDrops[i].y - j) >= 0; j++) {
-      int y = torrentDrops[i].y - j;
-      if (y < MATRIX_HEIGHT && !isInTextArea(torrentDrops[i].x, y)) {
-        uint8_t intensity = 255 - (j * 60);  // Faster fade for smaller drops
-        if (intensity < 80) intensity = 80;
-        uint16_t color = scaledEffectColor565(intensity/2, intensity/2, intensity);  // Light blue/white torrent
-        matrix.drawPixel(torrentDrops[i].x, y, color);
-      }
-    }
-  }
-}
-
-/**
- * Initialize stars
- */
-void initializeStars() {
-  for (int i = 0; i < NUM_STARS; i++) {
-    stars[i].x = random(0, MATRIX_WIDTH * 100) / 100.0f;  // Sub-pixel precision
-    stars[i].y = random(0, MATRIX_HEIGHT * 100) / 100.0f;
-    stars[i].brightness = random(50, 255);
-    stars[i].twinkleState = random(0, 2);
-    stars[i].lastTwinkle = millis() + random(0, 2000);
-  }
-}
-
-/**
- * Initialize shooting stars
- */
-void initializeShootingStars() {
-  for (int i = 0; i < NUM_SHOOTING_STARS; i++) {
-    shootingStars[i].active = false;
-  }
-  lastShootingStarTime = millis();
-}
-
-/**
- * Spawn a new shooting star
- */
-void spawnShootingStar(int index) {
-  // Start from random position at the top or left edge
-  if (random(0, 2) == 0) {
-    // Start from top edge
-    shootingStars[index].x = random(0, MATRIX_WIDTH);
-    shootingStars[index].y = 0;
-  } else {
-    // Start from left edge
-    shootingStars[index].x = 0;
-    shootingStars[index].y = random(0, MATRIX_HEIGHT);
-  }
-  
-  // Diagonal movement (northwest to southeast)
-  shootingStars[index].speedX = SHOOTING_STAR_SPEED;
-  shootingStars[index].speedY = SHOOTING_STAR_SPEED * 0.7f;  // Slightly less vertical speed
-  shootingStars[index].active = true;
-  shootingStars[index].trailLength = SHOOTING_STAR_TRAIL_LENGTH;
-  shootingStars[index].spawnTime = millis();
-}
-
-/**
- * Update drifting starfield effect - stars slowly drift across the sky
- */
-void updateStars() {
-  uint32_t currentTime = millis();
-  static uint32_t lastDriftUpdate = 0;
-  static float globalDriftX = 0.005f;  // Slowed down to 10% of original speed (0.05f -> 0.005f)
-  static float globalDriftY = 0.003f;  // Slowed down to 10% of original speed (0.03f -> 0.003f)
-  
-  // Slow drift movement - update every 150ms
-  if (currentTime - lastDriftUpdate > 150) {
-    // Use FIXED drift values for perfect group movement (no random variations during movement)
-    float groupDriftX = globalDriftX;  // Completely fixed drift
-    float groupDriftY = globalDriftY;
-    
-    // Apply the SAME movement to all stars as a unified group
-    for (int i = 0; i < NUM_STARS; i++) {
-      // All stars move together with the exact same drift
-      stars[i].x += groupDriftX;
-      stars[i].y += groupDriftY;
-      
-      // If star has drifted off the right or bottom edge, respawn it on the opposite edges
-      if (stars[i].x > MATRIX_WIDTH + 2 || stars[i].y > MATRIX_HEIGHT + 2) {
-        // Spawn from left or top edges with proper coordinates
-        if (stars[i].x > MATRIX_WIDTH + 2) {
-          // Respawn on left side
-          stars[i].x = -1.5f;  // Fixed spawn position to avoid random() calls
-          stars[i].y = (i * MATRIX_HEIGHT / NUM_STARS) + (i % 3 - 1);  // Distribute evenly with slight offset
-        }
-        if (stars[i].y > MATRIX_HEIGHT + 2) {
-          // Respawn on top side
-          stars[i].y = -1.5f;  // Fixed spawn position
-          stars[i].x = (i * MATRIX_WIDTH / NUM_STARS) + (i % 3 - 1);   // Distribute evenly with slight offset
-        }
-        // Keep existing brightness to avoid random() call
-      }
-    }
-    lastDriftUpdate = currentTime;
-  }
-  
-  // Shooting star management
-  // Check if it's time to spawn a shooting star (every 5-10 minutes)
-  if (currentTime - lastShootingStarTime > random(300000, 600000)) {  // 5-10 minutes
-  // if (currentTime - lastShootingStarTime > 5000) {  // TESTING: 5 seconds (COMMENTED OUT)
-    // Find an inactive shooting star to spawn
-    for (int i = 0; i < NUM_SHOOTING_STARS; i++) {
-      if (!shootingStars[i].active) {
-        spawnShootingStar(i);
-        lastShootingStarTime = currentTime;
-        
-        // 30% chance for a second shooting star
-        if (random(0, 100) < 30) {
-          waitingForSecondStar = true;
-          secondStarTimer = currentTime + random(2000, 5000);  // 2-5 seconds later
-        }
-        break;
-      }
-    }
-  }
-  
-  // Check for second shooting star
-  if (waitingForSecondStar && currentTime >= secondStarTimer) {
-    for (int i = 0; i < NUM_SHOOTING_STARS; i++) {
-      if (!shootingStars[i].active) {
-        spawnShootingStar(i);
-        waitingForSecondStar = false;
-        break;
-      }
-    }
-  }
-  
-  // Update shooting stars
-  for (int i = 0; i < NUM_SHOOTING_STARS; i++) {
-    if (shootingStars[i].active) {
-      shootingStars[i].x += shootingStars[i].speedX;
-      shootingStars[i].y += shootingStars[i].speedY;
-      
-      // Deactivate if off screen
-      if (shootingStars[i].x > MATRIX_WIDTH + 10 || shootingStars[i].y > MATRIX_HEIGHT + 10) {
-        shootingStars[i].active = false;
-      } else {
-        // Draw shooting star trail
-        for (int t = 0; t < shootingStars[i].trailLength; t++) {
-          int trailX = (int)(shootingStars[i].x - t * shootingStars[i].speedX * 0.5f);
-          int trailY = (int)(shootingStars[i].y - t * shootingStars[i].speedY * 0.5f);
-          
-          if (trailX >= 0 && trailX < MATRIX_WIDTH && trailY >= 0 && trailY < MATRIX_HEIGHT) {
-            if (!isInTextArea(trailX, trailY)) {
-              uint8_t brightness = 255 - (t * 32);  // Fade trail
-              uint16_t color = scaledEffectColor565(brightness, brightness, brightness);
-              matrix.drawPixel(trailX, trailY, color);
-            }
-          }
-        }
-      }
-    }
-  }
-  
-  for (int i = 0; i < NUM_STARS; i++) {
-    // Handle twinkling
-    if (currentTime - stars[i].lastTwinkle > random(800, 2000)) {
-      stars[i].twinkleState = !stars[i].twinkleState;
-      stars[i].lastTwinkle = currentTime;
-    }
-    
-    // Draw star if on screen and not in text area
-    int pixelX = (int)round(stars[i].x);
-    int pixelY = (int)round(stars[i].y);
-    
-    if (pixelX >= 0 && pixelX < MATRIX_WIDTH && pixelY >= 0 && pixelY < MATRIX_HEIGHT) {
-      if (!isInTextArea(pixelX, pixelY)) {
-        uint8_t brightness = stars[i].twinkleState ? stars[i].brightness : stars[i].brightness / 3;
-        uint16_t color = scaledEffectColor565(brightness, brightness, brightness);
-        matrix.drawPixel(pixelX, pixelY, color);
-      }
-    }
-  }
-}
-
-/**
- * Initialize sparkles
- */
-void initializeSparkles() {
-  for (int i = 0; i < NUM_SPARKLES; i++) {
-    sparkles[i].x = random(0, MATRIX_WIDTH);
-    sparkles[i].y = random(0, MATRIX_HEIGHT);
-    sparkles[i].brightness = 0;
-    sparkles[i].color = randomVividColor();  // Assign random vivid color
-    sparkles[i].startTime = millis() + random(0, 3000);
-    sparkles[i].duration = random(400, SPARKLE_DURATION);
-  }
-}
-
-/**
- * Update sparkle effect
- */
-void updateSparkles() {
-  uint32_t currentTime = millis();
-  
-  for (int i = 0; i < NUM_SPARKLES; i++) {
-    uint32_t elapsed = currentTime - sparkles[i].startTime;
-    
-    if (elapsed < sparkles[i].duration) {
-      // Calculate sparkle brightness (fade in and out)
-      float progress = (float)elapsed / sparkles[i].duration;
-      float brightness = sin(progress * PI) * 255;
-      sparkles[i].brightness = (uint8_t)brightness;
-      
-      // Draw sparkle if not in text area
-      if (!isInTextArea(sparkles[i].x, sparkles[i].y)) {
-        // Scale the stored color by brightness and apply global brightness
-        uint8_t r = ((sparkles[i].color >> 11) & 0x1F) * sparkles[i].brightness / 255;
-        uint8_t g = ((sparkles[i].color >> 5) & 0x3F) * sparkles[i].brightness / 255;
-        uint8_t b = (sparkles[i].color & 0x1F) * sparkles[i].brightness / 255;
-        uint16_t scaledColor = scaledEffectColor565(r << 3, g << 2, b << 3);
-        matrix.drawPixel(sparkles[i].x, sparkles[i].y, scaledColor);
-      }
-    } else {
-      // Reset sparkle to new position with new color
-      sparkles[i].x = random(0, MATRIX_WIDTH);
-      sparkles[i].y = random(0, MATRIX_HEIGHT);
-      sparkles[i].color = randomVividColor();  // New random color
-      sparkles[i].startTime = currentTime + random(0, 2000);
-      sparkles[i].duration = random(400, SPARKLE_DURATION);
-    }
-  }
-}
-
-/**
- * Initialize fireworks
- */
-void initializeFireworks() {
-  for (int i = 0; i < NUM_FIREWORKS; i++) {
-    fireworks[i].active = false;
-    fireworks[i].exploded = false;
-    fireworks[i].startTime = millis() + random(0, 3000);  // Stagger initial launches
-  }
-}
-
-/**
- * Update fireworks effect
- */
-void updateFireworks() {
-  uint32_t currentTime = millis();
-  
-  for (int i = 0; i < NUM_FIREWORKS; i++) {
-    if (!fireworks[i].active) {
-      // Launch new firework randomly
-      if (currentTime > fireworks[i].startTime) {
-        // Choose launch and explosion position to avoid text area
-        int textCenterX = MATRIX_WIDTH / 2;
-        int textCenterY = getCenteredY(settings.getTextSize());
-        
-        // Launch from sides or center, but explode away from text
-        if (random(0, 2) == 0) {
-          // Launch from left side, explode on left
-          fireworks[i].x = random(5, MATRIX_WIDTH / 3);
-        } else {
-          // Launch from right side, explode on right  
-          fireworks[i].x = random(2 * MATRIX_WIDTH / 3, MATRIX_WIDTH - 5);
-        }
-        
-        fireworks[i].y = MATRIX_HEIGHT;  // Start from bottom
-        fireworks[i].color = randomVividColor();  // Color for explosion
-        fireworks[i].active = true;
-        fireworks[i].exploded = false;
-        fireworks[i].startTime = currentTime;
-      }
-    } else if (!fireworks[i].exploded) {
-      // Rising phase - draw white rocket moving up
-      uint32_t elapsed = currentTime - fireworks[i].startTime;
-      float progress = elapsed / 800.0f;  // 800ms rise time
-      
-      if (progress >= 1.0f) {
-        // Time to explode - initialize particles
-        fireworks[i].exploded = true;
-        fireworks[i].startTime = currentTime;
-        
-        for (int j = 0; j < FIREWORK_PARTICLES; j++) {
-          float angle = (j * 2.0f * PI) / FIREWORK_PARTICLES;
-          float speed = random(100, 250) / 100.0f;  // Increased speed for more dramatic spread
-          fireworks[i].vx[j] = cos(angle) * speed;
-          fireworks[i].vy[j] = sin(angle) * speed;
-        }
-      } else {
-        // Draw rising white rocket
-        int rocketY = fireworks[i].y - (progress * (MATRIX_HEIGHT / 2));
-        if (rocketY >= 0 && rocketY < MATRIX_HEIGHT && !isInTextArea(fireworks[i].x, rocketY)) {
-          // White rocket trail
-          uint16_t whiteColor = applyEffectBrightness(matrix.color565(255, 255, 255));
-          matrix.drawPixel(fireworks[i].x, rocketY, whiteColor);
-        }
-      }
-    } else {
-      // Explosion phase - draw expanding particles
-      uint32_t elapsed = currentTime - fireworks[i].startTime;
-      float progress = elapsed / 1000.0f;  // Reduced from 1500ms to 1000ms for faster fade
-      
-      if (progress >= 1.0f) {
-        // Firework finished, reset for next launch
-        fireworks[i].active = false;
-        fireworks[i].startTime = currentTime + random(2000, 5000);  // Wait 2-5 seconds
-      } else {
-        // Draw explosion particles with more dramatic spread
-        float centerX = fireworks[i].x;
-        float centerY = fireworks[i].y - (MATRIX_HEIGHT / 2);
-        
-        for (int j = 0; j < FIREWORK_PARTICLES; j++) {
-          // Increased spread speed and reduced gravity effect
-          float px = centerX + (fireworks[i].vx[j] * elapsed / 50.0f);  // Increased from 100 to 50 for faster spread
-          float py = centerY + (fireworks[i].vy[j] * elapsed / 50.0f) + (0.1f * elapsed * elapsed / 10000.0f);  // Much reduced gravity
-          
-          if (px >= 0 && px < MATRIX_WIDTH && py >= 0 && py < MATRIX_HEIGHT) {
-            if (!isInTextArea(px, py)) {
-              // More dramatic fade - particles disappear while still in sky
-              float fade = (1.0f - progress) * (1.0f - progress);  // Exponential fade for more dramatic effect
-              if (fade > 0.1f) {  // Only draw if fade is significant
-                uint16_t fadedColor = applyEffectBrightness(scaleBrightness(fireworks[i].color, fade));
-                matrix.drawPixel((int)px, (int)py, fadedColor);
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-/**
- * Initialize Tron trails
- */
-void initializeTron() {
-  for (int i = 0; i < NUM_TRON_TRAILS; i++) {
-    tronTrails[i].active = false;
-    tronTrails[i].currentLength = 0;
-    tronTrails[i].lastMove = millis() + random(0, 2000);  // Stagger initial starts
-  }
-}
-
-/**
- * Update Tron effect - creates light trails across the grid
- */
-void updateTron() {
-  uint32_t currentTime = millis();
-  
-  for (int i = 0; i < NUM_TRON_TRAILS; i++) {
-    if (!tronTrails[i].active) {
-      // Start new trail
-      if (currentTime > tronTrails[i].lastMove) {
-        tronTrails[i].active = true;
-        tronTrails[i].currentLength = 0;
-        tronTrails[i].speed = random(TRON_MIN_SPEED, TRON_MAX_SPEED + 1);
-        tronTrails[i].direction = random(0, 4);  // 0=right, 1=down, 2=left, 3=up
-        
-        // Choose starting position based on direction
-        switch (tronTrails[i].direction) {
-          case 0: // Moving right
-            tronTrails[i].x = 0;
-            tronTrails[i].y = random(0, MATRIX_HEIGHT);
-            break;
-          case 1: // Moving down
-            tronTrails[i].x = random(0, MATRIX_WIDTH);
-            tronTrails[i].y = 0;
-            break;
-          case 2: // Moving left
-            tronTrails[i].x = MATRIX_WIDTH - 1;
-            tronTrails[i].y = random(0, MATRIX_HEIGHT);
-            break;
-          case 3: // Moving up
-            tronTrails[i].x = random(0, MATRIX_WIDTH);
-            tronTrails[i].y = MATRIX_HEIGHT - 1;
-            break;
-        }
-        
-        // Choose Tron-like colors (cyan, blue, white)
-        int colorChoice = random(0, 3);
-        switch (colorChoice) {
-          case 0: tronTrails[i].color = matrix.color565(0, 255, 255); break;    // Cyan
-          case 1: tronTrails[i].color = matrix.color565(0, 150, 255); break;   // Light blue
-          case 2: tronTrails[i].color = matrix.color565(255, 255, 255); break; // White
-        }
-        
-        // Add starting position to trail
-        tronTrails[i].trailPositions[0][0] = tronTrails[i].x;
-        tronTrails[i].trailPositions[0][1] = tronTrails[i].y;
-        tronTrails[i].currentLength = 1;
-        tronTrails[i].lastMove = currentTime;
-      }
-    } else {
-      // Update active trail
-      if (currentTime - tronTrails[i].lastMove >= tronTrails[i].speed) {
-        // Move trail head
-        switch (tronTrails[i].direction) {
-          case 0: tronTrails[i].x++; break; // Right
-          case 1: tronTrails[i].y++; break; // Down
-          case 2: tronTrails[i].x--; break; // Left
-          case 3: tronTrails[i].y--; break; // Up
-        }
-        
-        // Add new head position to trail if on screen
-        if (tronTrails[i].x >= 0 && tronTrails[i].x < MATRIX_WIDTH && 
-            tronTrails[i].y >= 0 && tronTrails[i].y < MATRIX_HEIGHT) {
-          
-          // Shift trail positions
-          if (tronTrails[i].currentLength >= TRON_MAX_LENGTH) {
-            // Remove tail
-            for (int j = 0; j < TRON_MAX_LENGTH - 1; j++) {
-              tronTrails[i].trailPositions[j][0] = tronTrails[i].trailPositions[j + 1][0];
-              tronTrails[i].trailPositions[j][1] = tronTrails[i].trailPositions[j + 1][1];
-            }
-            tronTrails[i].currentLength = TRON_MAX_LENGTH - 1;
-          }
-          
-          // Add new head
-          tronTrails[i].trailPositions[tronTrails[i].currentLength][0] = tronTrails[i].x;
-          tronTrails[i].trailPositions[tronTrails[i].currentLength][1] = tronTrails[i].y;
-          tronTrails[i].currentLength++;
-        }
-        
-        // Draw the trail
-        for (int j = 0; j < tronTrails[i].currentLength; j++) {
-          uint8_t segX = tronTrails[i].trailPositions[j][0];
-          uint8_t segY = tronTrails[i].trailPositions[j][1];
-          
-          if (segX < MATRIX_WIDTH && segY < MATRIX_HEIGHT) {
-            if (!isInTextArea(segX, segY)) {
-              // Fade from dim at tail to bright at head
-              float brightness = (float)(j + 1) / tronTrails[i].currentLength;
-              uint16_t fadedColor = applyEffectBrightness(scaleBrightness(tronTrails[i].color, brightness));
-              matrix.drawPixel(segX, segY, fadedColor);
-            }
-          }
-        }
-        
-        // Check if trail is off screen
-        if (tronTrails[i].x < -5 || tronTrails[i].x > MATRIX_WIDTH + 5 ||
-            tronTrails[i].y < -5 || tronTrails[i].y > MATRIX_HEIGHT + 5) {
-          tronTrails[i].active = false;
-          tronTrails[i].lastMove = currentTime + random(1000, 3000);  // Wait before next trail
-        } else {
-          tronTrails[i].lastMove = currentTime;
-        }
-      } else {
-        // Still draw existing trail even when not moving
-        for (int j = 0; j < tronTrails[i].currentLength; j++) {
-          uint8_t segX = tronTrails[i].trailPositions[j][0];
-          uint8_t segY = tronTrails[i].trailPositions[j][1];
-          
-          if (segX < MATRIX_WIDTH && segY < MATRIX_HEIGHT) {
-            if (!isInTextArea(segX, segY)) {
-              // Fade from dim at tail to bright at head
-              float brightness = (float)(j + 1) / tronTrails[i].currentLength;
-              uint16_t fadedColor = applyEffectBrightness(scaleBrightness(tronTrails[i].color, brightness));
-              matrix.drawPixel(segX, segY, fadedColor);
-            }
-          }
-        }
-      }
-    }
-  }
-}
 
 /**
  * Draw black background rectangles around the text areas
@@ -1246,13 +483,13 @@ void drawTextBackground() {
   
   // Draw background for main clock
   int x1, y1, x2, y2;
-  getTimeDisplayBounds(x1, y1, x2, y2);
-  matrix.fillRect(x1, y1, x2 - x1 + 1, y2 - y1 + 1, 0x0000);
+  display.getTimeDisplayBounds(x1, y1, x2, y2);
+  display.fillRect(x1, y1, x2 - x1 + 1, y2 - y1 + 1, 0x0000);
   
   // Draw background for AM/PM if in 12-hour format
   if (!settings.getUse24HourFormat()) {
-    getAMPMDisplayBounds(x1, y1, x2, y2);
-    matrix.fillRect(x1, y1, x2 - x1 + 1, y2 - y1 + 1, 0x0000);
+    display.getAMPMDisplayBounds(x1, y1, x2, y2);
+    display.fillRect(x1, y1, x2 - x1 + 1, y2 - y1 + 1, 0x0000);
   }
 }
 
@@ -1260,52 +497,8 @@ void drawTextBackground() {
  * Render background effects
  */
 void renderEffects() {
-  switch (settings.getEffectMode()) {
-    case EFFECT_CONFETTI:
-      updateConfetti();
-      drawTextBackground();  // Draw black background behind text
-      break;
-      
-    case EFFECT_ACID:
-      updateMatrixRain();  // Green acid rain
-      drawTextBackground();
-      break;
-      
-    case EFFECT_RAIN:
-      updateRain();  // Blue rain
-      drawTextBackground();
-      break;
-      
-    case EFFECT_TORRENT:
-      updateTorrent();  // Heavy rain with many small drops
-      drawTextBackground();
-      break;
-      
-    case EFFECT_STARS:
-      updateStars();
-      drawTextBackground();
-      break;
-      
-    case EFFECT_SPARKLES:
-      updateSparkles();
-      drawTextBackground();
-      break;
-      
-    case EFFECT_FIREWORKS:
-      updateFireworks();
-      drawTextBackground();
-      break;
-      
-    case EFFECT_TRON:
-      updateTron();
-      drawTextBackground();
-      break;
-      
-    case EFFECT_OFF:
-    default:
-      // No effects, just black background
-      break;
-  }
+  effects.updateEffects();
+  drawTextBackground();  // Draw black background behind text for all effects except OFF
 }
 
 // ===============================================
@@ -1888,16 +1081,8 @@ void initializeSystem() {
     rtc.adjust(DateTime(F(__DATE__), F(__TIME__)));
   }
   
-  // Initialize effects
-  initializeConfetti();
-  initializeMatrixRain();  // Acid rain (green)
-  initializeRain();        // Blue rain
-  initializeTorrent();     // Heavy rain
-  initializeStars();
-  initializeShootingStars();
-  initializeSparkles();
-  initializeFireworks();
-  initializeTron();         // Tron light trails
+  // Initialize effects engine
+  effects.begin();
   
   Serial.println("Setup complete!");
 }
